requirement.txt
langchain
langchain-openai
langsmith
python-dotenv
redis
pymongo